Make number-y columns into actual type:number columns that allow greater-than/less-than filtering
