+++ plugins/linear-genome-view/src/LinearGenomeView/util.ts
@@ -1,6 +1,7 @@
-import { parseLocString } from '@jbrowse/core/util'
+import { assembleLocString, parseLocString } from '@jbrowse/core/util'
 
 import type { AssemblyManager, ParsedLocString } from '@jbrowse/core/util'
+import type { BaseBlock } from '@jbrowse/core/util/blockTypes'
 
 /**
  * Given a scale ( bp/px ) and minimum distances (px) between major and minor
@@ -172,5 +173,24 @@ export function parseLocStrings(
       ]
     }
     throw e
+  }
+}
+
+export function calculateVisibleLocStrings(contentBlocks: BaseBlock[]) {
+  if (!contentBlocks.length) {
+    return ''
+  } else {
+    const isSingleAssemblyName = contentBlocks.every(
+      b => b.assemblyName === contentBlocks[0]!.assemblyName,
+    )
+    const locs = contentBlocks.map(block =>
+      assembleLocString({
+        ...block,
+        start: Math.round(block.start),
+        end: Math.round(block.end),
+        assemblyName: isSingleAssemblyName ? undefined : block.assemblyName,
+      }),
+    )
+    return locs.join(' ')
   }
 }
