+++ plugins/spreadsheet-view/src/SpreadsheetView/components/SpreadsheetDataGrid.tsx
@@ -1,4 +1,3 @@
-import { LoadingEllipses } from '@jbrowse/core/ui'
 import { DataGrid, GridToolbar, useGridApiRef } from '@mui/x-data-grid'
 import { observer } from 'mobx-react'
 
@@ -41,9 +40,7 @@ const SpreadsheetDataGrid = observer(function ({
       rows={rows}
       columns={dataGridColumns}
     />
-  ) : (
-    <LoadingEllipses variant="h6" />
-  )
+  ) : null
 })
 
 export default SpreadsheetDataGrid
